Parse as UTF-8 before sending and after receiving
The change uses/returns UTF-8 when using Invoke-WebRequest, both for the body
being sent (PUT or POST methods) and for the content received from the
server.

As the Isilon API don't return character encoding in it's headers (but
seems to be using UTF-8), the response is interpreted as something else
(ISO-8859-1? Or perhaps set by the system invoking). See [article][1].

[1]: https://stackoverflow.com/a/38699430/2670584